+++ pirate_games/src/pages/PirateGames/PirateGamesWrapper.tsx
@@ -1,0 +1,17 @@
+import { FC } from "react";
+import { PirateGamesProvider } from "./PirateGamesContext";
+import { PaginaInicial } from "./pages/PaginaInicial";
+
+const PirateGamesWrapper: FC = () => {
+
+    return (
+        <>
+        <PirateGamesProvider>
+            <PaginaInicial/>
+        </PirateGamesProvider>
+        </>
+    )
+
+}
+
+export { PirateGamesWrapper }
