+++ pirate_games/src/pages/PirateGames/components/Carousel/Carousel.tsx
@@ -5,27 +5,13 @@ import "./Carousel.css";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Autoplay, Pagination, Navigation } from "swiper/modules";
 import { Imagen } from "../../../../interfaces/General";
+import { match } from "assert";
 
-const noticias: Array<Imagen> = [
-    {
-        id: 1,
-        imgUrl: "./img/black banner.jpg",
-    },
-    {
-        id: 2,
-        imgUrl: "./img/DBZ Budokai Tenkaichi 4 banner.jpg",
-    },
-    {
-        id: 3,
-        imgUrl: "./img/Dios de la guerra 3 banner.jpg",
-    },
-    {
-        id: 4,
-        imgUrl: "./img/Efootball 2024 banner.jpg",
-    },
-];
+type Noticias = {
+    noticias: Array<Imagen>;
+}
 
-const Carousel = () => {
+const Carousel = ({noticias}: Noticias) => {
     return (
         <>
             <Swiper
@@ -48,7 +34,7 @@ const Carousel = () => {
                 {
                     noticias.map((item) => {
                         return (
-                            <SwiperSlide className="">
+                            <SwiperSlide key={Math.random()}>
                                 <img className='imagen-noticia' src={item.imgUrl} />
                             </SwiperSlide>
                         );
