+++ pirate_games/src/pages/PirateGames/components/CartaDeJuego/CaratulaJuego.css
@@ -13,6 +13,7 @@
     border: 3px solid transparent;
     background-clip: padding-box;
     text-align: center;
+    cursor: pointer;
     color: #f1f3f3;
     background-image: linear-gradient(to bottom right,
             #1b2028,
@@ -74,4 +75,70 @@
     margin-top: -0.7rem;
     margin-bottom: -1rem;
     color: #7d8396;
+}
+
+@media screen and (max-width: 935px) {
+
+    .profile {
+        padding: 1rem;
+        width: 170px;
+        height: 310px;
+    }
+    
+    .profile:active {
+        color: #fdff00;
+    }
+    
+    .profile-image {
+        width: 150px;
+        height: 220px;
+    }
+    
+    .profile-username {
+        font-size: 14px;
+        padding: 0.3rem;
+    }
+    
+    .profile-user-handle {
+        font-size: 12px;
+        color: #7d8396;
+    }
+
+}
+
+/*RESPONSIVE PARA TELEFONO*/
+
+@media screen and (max-width: 620px) {
+
+    .profile {
+        padding: 1rem;
+        width: 160px;
+        height: 300px;
+    }
+    
+    .profile-after {
+        border-radius: 16px;
+        background-image: linear-gradient(300deg,
+                #2f312c 30%,
+                #fdff00 100%);
+    }
+    
+    .profile:active {
+        color: #fdff00;
+    }
+    
+    .profile-image {
+        width: 140px;
+        height: 210px;
+    }
+    
+    .profile-username {
+        font-size: 12px;
+        padding: 0.3rem;
+    }
+    
+    .profile-user-handle {
+        font-size: 12px;
+        color: #7d8396;
+    }
 }
