+++ pirate_games/src/pages/PirateGames/pages/InicioPirateGames/InicioPirateGames.tsx
@@ -1,54 +1,15 @@
-import { Juego } from "../../../../interfaces/General";
 import { Carousel } from "../../components/Carousel/Carousel";
 import { CaratulaJuego } from "../../components/CartaDeJuego/CaratulaJuego";
+import { juegos, noticias } from "../TS/GeneralApi";
 import "./InicioPirateGames.css";
 import { FC } from "react"
-
-const juegos: Array<Juego> = [
-    {
-        id: 1,
-        imagen: "https://img.asmedia.epimg.net/resizer/lqsy7mJHPWIbj45aTSL7gwZY4U8=/360x0/cloudfront-eu-central-1.images.arcpublishing.com/diarioas/KJSGVNUE3ZJSFLZ5MUJVKHHILY.jpg",
-        nombre: "Grand Theft Auto: San Andreas",
-        region: "Ntfs-Pal",
-    },
-    {
-        id: 2,
-        imagen: "./img/Need for Speed Most Wanted.jpg",
-        nombre: "Need for Speed Most Wanted",
-        region: "Ntfs-Pal",
-    },
-    {
-        id: 3,
-        imagen: "./img/Naruto Ultimate Ninja 3.jpg",
-        nombre: "Naruto Ultimate Ninja 3",
-        region: "Ntfs-Pal",
-    },
-    {
-        id: 4,
-        imagen: "./img/PES 2020 3.0 FINAL.jpg",
-        nombre: "PES 2020 3.0 FINAL",
-        region: "Ntfs-Pal",
-    },
-    {
-        id: 5,
-        imagen: "./img/DBZ Budokai Tenkaichi 3.jpg",
-        nombre: "DBZ Budokai Tenkaichi 3",
-        region: "Ntfs-Pal",
-    },
-    {
-        id: 6,
-        imagen: "./img/Bully Canis Canem Edit.jpg",
-        nombre: "Bully Canis Canem Edit",
-        region: "Ntfs-Pal",
-    },
-];
 
 const InicioPirateGames: FC = () => {
     return (
         <>
             <div className="body-cartas">
                 <div className="back-carrusel">
-                    <Carousel />
+                    <Carousel noticias={noticias} />
                 </div>
                 <div className="back-cartas">
                     <br />
@@ -65,7 +26,7 @@ const InicioPirateGames: FC = () => {
                         {
                             juegos.map((item) => {
                                 return (
-                                    <div key={item.id} className="carta">
+                                    <div key={item.id} className="carta" onClick={(e) => (console.log(item.id))}>
                                         <CaratulaJuego
                                             imagen={item.imagen}
                                             nombre={item.nombre}
