+++ pirate_games/src/pages/PirateGames/pages/TS/GeneralApi.ts
@@ -1,0 +1,65 @@
+import { Imagen, Juego } from "../../../../interfaces/General";
+
+export const juegos: Array<Juego> = [
+    {
+        id: 1,
+        imagen: "https://img.asmedia.epimg.net/resizer/lqsy7mJHPWIbj45aTSL7gwZY4U8=/360x0/cloudfront-eu-central-1.images.arcpublishing.com/diarioas/KJSGVNUE3ZJSFLZ5MUJVKHHILY.jpg",
+        nombre: "Grand Theft Auto: San Andreas",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 2,
+        imagen: "./img/Need for Speed Most Wanted.jpg",
+        nombre: "Need for Speed Most Wanted",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 3,
+        imagen: "./img/Naruto Ultimate Ninja 3.jpg",
+        nombre: "Naruto Ultimate Ninja 3",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 4,
+        imagen: "./img/PES 2020 3.0 FINAL.jpg",
+        nombre: "PES 2020 3.0 FINAL",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 5,
+        imagen: "./img/DBZ Budokai Tenkaichi 3.jpg",
+        nombre: "DBZ Budokai Tenkaichi 3",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 6,
+        imagen: "./img/Bully Canis Canem Edit.jpg",
+        nombre: "Bully Canis Canem Edit",
+        region: "Ntfs-Pal",
+    },
+    {
+        id: 7,
+        imagen: "./img/Bully Canis Canem Edit.jpg",
+        nombre: "Bully Canis Canem Edit",
+        region: "Ntfs-Pal",
+    },
+];
+
+export const noticias: Array<Imagen> = [
+    {
+        id: 1,
+        imgUrl: "./img/black banner.jpg",
+    },
+    {
+        id: 2,
+        imgUrl: "./img/DBZ Budokai Tenkaichi 4 banner.jpg",
+    },
+    {
+        id: 3,
+        imgUrl: "./img/Dios de la guerra 3 banner.jpg",
+    },
+    {
+        id: 4,
+        imgUrl: "./img/Efootball 2024 banner.jpg",
+    },
+];
